testing lfs tracking and small code cleaning

+++ smart_simulation/database/db_utils.py
@@ -1,6 +1,7 @@
 import logging
 import pathlib
 import sqlite3
+from functools import lru_cache
 
 import pandas as pd
 
@@ -22,6 +23,7 @@ PACKAGE_PATH = pathlib.Path(package_dir)
 LOCAL_PROD_DB = PACKAGE_PATH / "smart_simulation" / "database" / "production.db"
 
 
+@lru_cache(maxsize=None)
 def connect(database: pathlib.PurePath = LOCAL_PROD_DB) -> sqlite3.connect:
     """
     Connect to a sqlite3 database and return the connection.
@@ -55,18 +57,17 @@ def weight_csv_to_db(
     """
     # TODO add in check for datetime format of index, and that weight column exists
     # TODO and create temp df with just date_time, scale_id, weight
-    connection = connect(database)
-    scale_df = pd.read_csv(csv_file, parse_dates=True, index_col=0)
-    scale_df.loc[:, "scale_id"] = truncated_uuid
-    scale_df.to_sql(
-        "weights",
-        con=connection,
-        if_exists="append",
-        index=True,
-        dtype="TEXT",
-        index_label="date_time",
-    )
-    connection.close()
+    with connect(database) as connection:
+        scale_df = pd.read_csv(csv_file, parse_dates=True, index_col=0)
+        scale_df.loc[:, "scale_id"] = truncated_uuid
+        scale_df.to_sql(
+            "weights",
+            con=connection,
+            if_exists="append",
+            index=True,
+            dtype="TEXT",
+            index_label="date_time",
+        )
 
 
 def delete_all_rows(table: str, database: pathlib.PurePath = LOCAL_PROD_DB):
@@ -76,9 +77,14 @@ def delete_all_rows(table: str, database: pathlib.PurePath = LOCAL_PROD_DB):
         table: Table to delete records from.
         database: Database to delete records from, defaulted to the local production database.
     """
-    connection = connect(database)
-    c = connection.cursor()
-    sql_statement = f"DELETE FROM {table};"
-    c.execute(sql_statement,)
-    connection.commit()
-    connection.close()
+    if not isinstance(table, str):
+        logging.exception(f"table must be of type: str. Received a {type(table)}.")
+        raise TypeError
+
+    with connect(database) as connection:
+        c = connection.cursor()
+        sql_statement = f"DELETE FROM {table};"
+        try:
+            c.execute(sql_statement,)
+        except Exception:
+            raise
